openstack_gerrit/requirements.txt
requests
numpy
paramiko
six
tflearn
pyyaml